Moved bridge into adapter
